+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Improvements
 - YEAR : disable canEdit on company field in year-form.
 - MRP : add a try catch block in generateAllProposals method.
+- TIMESHEET : Throw an alert on timesheet validation.
 
 ## Bug Fixes
 - STOCK MOVE : fixed issue on printing where origin sale order's reference number wouldn't be displayed